Fix zookeeper image name + minor tag update
Shouldn't the zookeeper image be strimzi/zookeeper instead of strimzi/kafka?

At the same time, considering it's a minor release increment, I'm also suggesting the tag push from 0.11.3-kafka-2.1.0 to 0.11.4-kafka-2.1.0. For the purpose of the exercise/guide, the minor version increase has no impact.

Last but not the least, a big thank you to the original authors of the quarkus.io guides. I'm really enjoying it as someone who's worked with spring/spring boot.